Take image push option from the outside
Right now we have `--dry-run` to identify the actions we need to build
inside a particular git range.

We want to actually try to build the image but the `push` option was
fixed to `true`. This commit adds a `--push` option that has to be
specified in order to push the image to a registry.

This follows the same behavior buildx has.

Signed-off-by: Gianluca Arbezzano <ciao@gianarb.it>